+++ hl_image_picker_platform_interface/lib/src/camera_options.dart
@@ -1,38 +1,51 @@
-import 'package:hl_image_picker_platform_interface/src/constants.dart';
+import 'package:hl_image_picker_platform_interface/hl_image_picker_platform_interface.dart';
 
 class HLCameraOptions {
+  const HLCameraOptions({
+    this.cameraType,
+    this.recordVideoMaxSecond,
+    this.isExportThumbnail,
+    this.compressQuality,
+    this.compressFormat,
+    this.thumbnailCompressQuality,
+    this.thumbnailCompressFormat,
+    this.maxSizeOutput,
+  });
+
   /// Specifies the type of camera to be used: [CameraType.video], [CameraType.image]
   ///
   /// Default: [CameraType.image]
-  CameraType? cameraType;
+  final CameraType? cameraType;
 
   /// The maximum duration (in seconds) for recorded video.
   ///
   /// Default: `60` seconds
-  int? recordVideoMaxSecond;
+  final int? recordVideoMaxSecond;
 
   /// Determines whether to export thumbnail for recorded video.
-  bool? isExportThumbnail;
+  final bool? isExportThumbnail;
+
+  /// The compression quality for selected image.
+  final double? compressQuality;
+
+  /// The image format for selected image: [CompressFormat.jpg], [CompressFormat.png].
+  ///
+  /// Default: [CompressFormat.jpg]
+  final CompressFormat? compressFormat;
 
   /// The compression quality for exported thumbnail.
   ///
   /// Min: `0.1` - Max: `1`
   ///
   /// Default: `0.9`
-  double? thumbnailCompressQuality;
+  final double? thumbnailCompressQuality;
 
   /// The image format for exported thumbnail: [CompressFormat.jpg], [CompressFormat.png].
-  /// 
+  ///
   /// Default: [CompressFormat.jpg]
-  CompressFormat? thumbnailCompressFormat;
+  final CompressFormat? thumbnailCompressFormat;
 
-  HLCameraOptions({
-    this.cameraType,
-    this.recordVideoMaxSecond,
-    this.isExportThumbnail,
-    this.thumbnailCompressQuality,
-    this.thumbnailCompressFormat,
-  });
+  final MaxSizeOutput? maxSizeOutput;
 
   Map<String, dynamic> toMap() {
     final result = <String, dynamic>{};
@@ -41,6 +54,9 @@ class HLCameraOptions {
     result['isExportThumbnail'] = isExportThumbnail;
     result['thumbnailCompressQuality'] = thumbnailCompressQuality;
     result['thumbnailCompressFormat'] = thumbnailCompressFormat;
+    result['cameraCompressQuality'] = compressQuality;
+    result['cameraCompressFormat'] = compressFormat;
+    result['cameraMaxSizeOutput'] = maxSizeOutput?.toMap();
     return result;
   }
 }