+++ hl_image_picker_platform_interface/lib/src/crop_options.dart
@@ -1,13 +1,14 @@
 import 'constants.dart';
 
 class CropAspectRatio {
-  final double ratioX;
-  final double ratioY;
   const CropAspectRatio({required this.ratioX, required this.ratioY})
       : assert(ratioX > 0.0 && ratioY > 0.0);
 
-  Map<dynamic, dynamic> toMap() {
-    final data = <dynamic, dynamic>{};
+  final double ratioX;
+  final double ratioY;
+
+  Map<String, dynamic> toMap() {
+    final data = <String, dynamic>{};
     data['ratioX'] = ratioX;
     data['ratioY'] = ratioY;
     return data;
@@ -15,13 +16,13 @@ class CropAspectRatio {
 }
 
 class MaxSizeOutput {
+  const MaxSizeOutput({required this.maxWidth, required this.maxHeight});
+
   final int maxWidth;
   final int maxHeight;
 
-  MaxSizeOutput({required this.maxWidth, required this.maxHeight});
-
-  Map<dynamic, dynamic> toMap() {
-    final data = <dynamic, dynamic>{};
+  Map<String, dynamic> toMap() {
+    final data = <String, dynamic>{};
     data['maxWidth'] = maxWidth;
     data['maxHeight'] = maxHeight;
     return data;
@@ -29,8 +30,17 @@ class MaxSizeOutput {
 }
 
 class HLCropOptions {
+  const HLCropOptions({
+    this.croppingStyle,
+    this.aspectRatio,
+    this.aspectRatioPresets,
+    this.compressQuality,
+    this.compressFormat,
+    this.maxSizeOutput,
+  });
+
   /// Specifies the desired aspect ratio for cropping.
-  CropAspectRatio? aspectRatio;
+  final CropAspectRatio? aspectRatio;
 
   /// Provides a set of predefined aspect ratio options for cropping.
   ///
@@ -45,46 +55,37 @@ class HLCropOptions {
   ///  ];
   /// ```
   ///
-  List<CropAspectRatioPreset>? aspectRatioPresets;
+  final List<CropAspectRatioPreset>? aspectRatioPresets;
 
   /// Determines the compression quality for the exported image.
   ///
   /// Min: `0.1` - Max: `1`
   ///
   /// Default: `0.9`
-  double? compressQuality;
+  final double? compressQuality;
 
   /// Specifies the image format for the exported image: [CompressFormat.jpg], [CompressFormat.png].
   ///
   /// Default: [CompressFormat.jpg]
-  CompressFormat? compressFormat;
+  final CompressFormat? compressFormat;
 
   /// Cropping style: [CroppingStyle.normal], [CroppingStyle.circular].
   ///
   /// Default: [CroppingStyle.normal]
-  CroppingStyle? croppingStyle;
+  final CroppingStyle? croppingStyle;
 
   /// Sets the maximum width and maximum height for the exported image.
-  MaxSizeOutput? maxSizeOutput;
-
-  HLCropOptions({
-    this.croppingStyle,
-    this.aspectRatio,
-    this.aspectRatioPresets,
-    this.compressQuality,
-    this.compressFormat,
-    this.maxSizeOutput,
-  });
+  final MaxSizeOutput? maxSizeOutput;
 
   Map<String, dynamic> toMap() {
     final result = <String, dynamic>{};
     result['aspectRatio'] = aspectRatio?.toMap();
     result['aspectRatioPresets'] =
         aspectRatioPresets?.map((x) => x.value).toList();
-    result['compressQuality'] = compressQuality;
-    result['compressFormat'] = compressFormat?.name;
+    result['cropCompressQuality'] = compressQuality;
+    result['cropCompressFormat'] = compressFormat?.name;
     result['croppingStyle'] = croppingStyle?.name;
-    result['maxSizeOutput'] = maxSizeOutput?.toMap();
+    result['cropMaxSizeOutput'] = maxSizeOutput?.toMap();
     return result;
   }
 }