+++ hl_image_picker_platform_interface/lib/src/picker_localized.dart
@@ -1,11 +1,17 @@
 class LocalizedImageCropper {
+  const LocalizedImageCropper({
+    this.cropDoneText,
+    this.cropCancelText,
+    this.cropTitleText,
+  });
+
   /// The text displayed on the "Done" button.
   ///
   /// Platform: **iOS**
   ///
   /// Default:
   /// **'Done'**
-  late final String? cropDoneText;
+  final String? cropDoneText;
 
   /// The text displayed on the "Done" button.
   ///
@@ -13,22 +19,17 @@ class LocalizedImageCropper {
   ///
   /// Default:
   /// **'Cancel'**
-  late final String? cropCancelText;
+  final String? cropCancelText;
 
   /// The title displayed in the crop image screen.
-  late final String? cropTitleText;
+  final String? cropTitleText;
 
-  LocalizedImageCropper({
-    this.cropDoneText,
-    this.cropCancelText,
-    this.cropTitleText,
-  });
-
-  LocalizedImageCropper.fromMap(Map<String, dynamic> map) {
-    cropTitleText = map['cropTitleText'];
-    cropDoneText = map['cropDoneText'];
-    cropCancelText = map['cropCancelText'];
-  }
+  factory LocalizedImageCropper.fromMap(Map<String, dynamic> map) =>
+      LocalizedImageCropper(
+        cropDoneText: map['cropDoneText'],
+        cropCancelText: map['cropCancelText'],
+        cropTitleText: map['cropTitleText'],
+      );
 
   Map<String, dynamic> toMap() {
     final data = <String, dynamic>{};
@@ -40,113 +41,7 @@ class LocalizedImageCropper {
 }
 
 class LocalizedImagePicker extends LocalizedImageCropper {
-  /// The error message displayed when the selected video exceeds the maximum duration.
-  ///
-  /// Default:
-  /// **'Exceeded maximum duration of the video'**
-  late final String? maxDurationErrorText;
-
-  /// The error message displayed when the selected video is below the minimum duration.
-  ///
-  /// Default:
-  /// **'The video is too short'**
-  late final String? minDurationErrorText;
-
-  /// The error message displayed when the selected file exceeds the maximum file size.
-  ///
-  /// Default:
-  /// **'Exceeded maximum file size'**
-  late final String? maxFileSizeErrorText;
-
-  /// The error message displayed when the selected file is below the minimum file size.
-  ///
-  /// Default:
-  /// **'The file size is too small'**
-  late final String? minFileSizeErrorText;
-
-  /// The error message displayed when the app doesn't have permission to access the album.
-  ///
-  /// Default:
-  /// **'No permission to access album'**
-  late final String? noAlbumPermissionText;
-
-  /// The error message displayed when the app doesn't have permission to access the camera.
-  ///
-  /// Default:
-  /// **'TNo permission to access camera'**
-  late final String? noCameraPermissionText;
-
-  /// The error message displayed when the maximum number of items is exceeded.
-  ///
-  /// Default:
-  /// **'Exceeded maximum number of selected items'**
-  late final String? maxSelectedAssetsErrorText;
-
-  /// The error message displayed when the minimum number of items is not met.
-  ///
-  /// Default:
-  /// **'Need to select at least {minSelectedAssets}'**
-  late final String? minSelectedAssetsErrorText;
-
-  /// The text displayed on the "Done" button.
-  ///
-  /// Default:
-  /// **'Done'**
-  late final String? doneText;
-
-  /// The error message displayed when the app doesn't have permission to record audio.
-  ///
-  /// Platform: **Android**
-  ///
-  /// Default:
-  /// **'No permission to record audio'**
-  late final String? noRecordAudioPermissionText;
-
-  /// The text displayed on the "OK" button.
-  ///
-  /// Default:
-  /// **'OK'**
-  late final String? okText;
-
-  /// The text displayed when the picker is in a loading state.
-  ///
-  /// Default:
-  /// **'Loading'**
-  late final String? loadingText;
-
-  /// The name for default album.
-  ///
-  /// Default:
-  /// **'Recents'**
-  late final String? defaultAlbumName;
-
-  /// The text displayed on the "Cancel" button.
-  ///
-  /// Default:
-  /// **'Cancel'**
-  late final String? cancelText;
-
-  /// The text displayed below `defaultAlbumName`.
-  ///
-  /// Platform: **iOS**
-  ///
-  /// Default:
-  /// **'Tap here to change'**
-  late final String? tapHereToChangeText;
-
-  /// The text displayed when no media is available.
-  ///
-  /// Default:
-  /// **'No media available'**
-  late final String? emptyMediaText;
-
-  /// The text displayed when no media is available.
-  ///
-  /// Default:
-  /// **'File type is not supported'**
-  late final String? gifErrorText;
-
-  LocalizedImagePicker({
+  const LocalizedImagePicker({
     this.maxDurationErrorText,
     this.minDurationErrorText,
     this.maxFileSizeErrorText,
@@ -168,6 +63,112 @@ class LocalizedImagePicker extends LocalizedImageCropper {
     super.cropCancelText,
     super.cropTitleText,
   });
+
+  /// The error message displayed when the selected video exceeds the maximum duration.
+  ///
+  /// Default:
+  /// **'Exceeded maximum duration of the video'**
+  final String? maxDurationErrorText;
+
+  /// The error message displayed when the selected video is below the minimum duration.
+  ///
+  /// Default:
+  /// **'The video is too short'**
+  final String? minDurationErrorText;
+
+  /// The error message displayed when the selected file exceeds the maximum file size.
+  ///
+  /// Default:
+  /// **'Exceeded maximum file size'**
+  final String? maxFileSizeErrorText;
+
+  /// The error message displayed when the selected file is below the minimum file size.
+  ///
+  /// Default:
+  /// **'The file size is too small'**
+  final String? minFileSizeErrorText;
+
+  /// The error message displayed when the app doesn't have permission to access the album.
+  ///
+  /// Default:
+  /// **'No permission to access album'**
+  final String? noAlbumPermissionText;
+
+  /// The error message displayed when the app doesn't have permission to access the camera.
+  ///
+  /// Default:
+  /// **'TNo permission to access camera'**
+  final String? noCameraPermissionText;
+
+  /// The error message displayed when the maximum number of items is exceeded.
+  ///
+  /// Default:
+  /// **'Exceeded maximum number of selected items'**
+  final String? maxSelectedAssetsErrorText;
+
+  /// The error message displayed when the minimum number of items is not met.
+  ///
+  /// Default:
+  /// **'Need to select at least {minSelectedAssets}'**
+  final String? minSelectedAssetsErrorText;
+
+  /// The text displayed on the "Done" button.
+  ///
+  /// Default:
+  /// **'Done'**
+  final String? doneText;
+
+  /// The error message displayed when the app doesn't have permission to record audio.
+  ///
+  /// Platform: **Android**
+  ///
+  /// Default:
+  /// **'No permission to record audio'**
+  final String? noRecordAudioPermissionText;
+
+  /// The text displayed on the "OK" button.
+  ///
+  /// Default:
+  /// **'OK'**
+  final String? okText;
+
+  /// The text displayed when the picker is in a loading state.
+  ///
+  /// Default:
+  /// **'Loading'**
+  final String? loadingText;
+
+  /// The name for default album.
+  ///
+  /// Default:
+  /// **'Recents'**
+  final String? defaultAlbumName;
+
+  /// The text displayed on the "Cancel" button.
+  ///
+  /// Default:
+  /// **'Cancel'**
+  final String? cancelText;
+
+  /// The text displayed below `defaultAlbumName`.
+  ///
+  /// Platform: **iOS**
+  ///
+  /// Default:
+  /// **'Tap here to change'**
+  final String? tapHereToChangeText;
+
+  /// The text displayed when no media is available.
+  ///
+  /// Default:
+  /// **'No media available'**
+  final String? emptyMediaText;
+
+  /// The text displayed when no media is available.
+  ///
+  /// Default:
+  /// **'File type is not supported'**
+  final String? gifErrorText;
 
   @override
   Map<String, dynamic> toMap() {