+++ hl_image_picker_platform_interface/lib/src/picker_options.dart
@@ -1,7 +1,9 @@
+import 'package:hl_image_picker_platform_interface/src/crop_options.dart';
+
 import 'constants.dart';
 
 class HLPickerOptions {
-  HLPickerOptions({
+  const HLPickerOptions({
     this.numberOfColumn,
     this.usedCameraButton,
     this.mediaType,
@@ -19,81 +21,98 @@ class HLPickerOptions {
     this.maxDuration,
     this.minDuration,
     this.isGif,
+    this.compressFormat,
+    this.compressQuality,
+    this.maxSizeOutput,
   });
 
   /// Type of media you want to select: [MediaType.image], [MediaType.video], [MediaType.all].
   ///
   /// Default: [MediaType.all]
-  MediaType? mediaType;
+  final MediaType? mediaType;
 
   /// The maximum number of items that can be selected.
   ///
   /// Default: 1
-  int? maxSelectedAssets;
+  final int? maxSelectedAssets;
 
   /// The minimum number of items that must be selected.
-  int? minSelectedAssets;
+  final int? minSelectedAssets;
 
   /// The maximum allowed file size for selected items.
-  double? maxFileSize;
+  final double? maxFileSize;
 
   /// The minimum allowed file size for selected items.
-  double? minFileSize;
+  final double? minFileSize;
 
   /// Enables or disables the preview feature.
   ///
   /// **Press** on **Android** / **Long Press** on **iOS**
-  bool? enablePreview;
+  final bool? enablePreview;
 
   /// Converts HEIC format images to JPEG format when selected.
   ///
   /// Platform: iOS
-  bool? convertHeicToJPG;
+  final bool? convertHeicToJPG;
 
   /// Converts Live Photos to JPEG format when selected.
   ///
   /// Platform: iOS
-  bool? convertLivePhotosToJPG;
+  final bool? convertLivePhotosToJPG;
 
   /// The maximum duration (in seconds) for recorded video.
   ///
   /// Default: `60` seconds
-  int? recordVideoMaxSecond;
+  final int? recordVideoMaxSecond;
 
   /// Determines whether to export thumbnail for selected videos.
   ///
   /// Default: `false`
-  bool? isExportThumbnail;
+  final bool? isExportThumbnail;
 
   /// The compression quality for exported thumbnails.
   ///
   /// Min: `0.1` - Max: `1`
   ///
   /// Default: `0.9`
-  double? thumbnailCompressQuality;
+  final double? thumbnailCompressQuality;
 
   /// The image format for exported thumbnails: [CompressFormat.jpg], [CompressFormat.png].
   ///
   /// Default: [CompressFormat.jpg]
-  CompressFormat? thumbnailCompressFormat;
+  final CompressFormat? thumbnailCompressFormat;
 
   /// The maximum duration (in seconds) for selected videos.
-  int? maxDuration;
+  final int? maxDuration;
 
   /// The minimum duration (in seconds) for selected videos.
-  int? minDuration;
+  final int? minDuration;
 
   /// The number of items displayed per row in the picker list.
   ///
   /// Default value: `3`
-  late final int? numberOfColumn;
+  final int? numberOfColumn;
 
   /// Determines whether to show the camera button in the picker list.
   ///
   /// Default value: `true`
-  late final bool? usedCameraButton;
+  final bool? usedCameraButton;
 
-  bool? isGif;
+  /// Enable gif selection
+  ///
+  /// Default: `false`
+  final bool? isGif;
+
+  /// The compression quality for compressed images.
+  final double? compressQuality;
+
+  /// The image format for compressed images: [CompressFormat.jpg], [CompressFormat.png].
+  /// 
+  /// Default: [CompressFormat.jpg]
+  final CompressFormat? compressFormat;
+
+  /// Sets the maximum width and maximum height for selected images.
+  final MaxSizeOutput? maxSizeOutput;
 
   Map<String, dynamic> toMap() {
     final result = <String, dynamic>{};
@@ -114,6 +133,9 @@ class HLPickerOptions {
     result['numberOfColumn'] = numberOfColumn;
     result['usedCameraButton'] = usedCameraButton;
     result['isGif'] = isGif;
+    result['compressQuality'] = compressQuality;
+    result['compressFormat'] = compressFormat;
+    result['maxSizeOutput'] = maxSizeOutput?.toMap();
     return result;
   }
 }